fixing bugs does a lot of good, even when there is no bug
Because then, you look at all of the code seeking for the little
mistake, and find a lot of little mistake that did not block you.

As much as we hate bugs, the bugs help us programming with care,
reminding us that we need to make things very clear for ourself
as we may have to debug our own mess at any time.